web-dashboard/web-backend/python_server/requirements.txt
aadict==0.2.3
altgraph==0.17.2
asset==0.6.13
astroid==2.11.5
async-timeout==4.0.2
backports.entry-points-selectable==1.1.1
bandit==1.7.4
bcrypt==3.2.0
bson==0.5.10
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
colorlog==6.6.0
Cython==0.29.28
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.3
dlib==19.24.0
dnspython==2.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg==1.4
filelock==3.4.0
fire==0.4.0
Flask==2.1.2
Flask-Cors==3.0.10
gitdb==4.0.9
GitPython==3.1.27
globre==0.1.5
idna==3.2
importlib-metadata==4.11.3
imutils==0.5.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
macholib==1.16
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
passwordmeter==0.1.8
pbr==5.9.0
Pillow==9.1.0
pipenv==2021.11.23
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.21
pyinstaller @ file:///Users/hritik/Desktop/ENGAGE/app/x/pyinstaller
pyinstaller-hooks-contrib==2022.4
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
randomname==0.1.5
redis==4.3.1
requests==2.26.0
scipy==1.8.0
six==1.16.0
smmap==5.0.0
stevedore==3.5.0
termcolor==1.1.0
tomli==2.0.1
tqdm==4.62.2
typing_extensions==4.2.0
urllib3==1.26.6
uuid==1.30
vidgear==0.2.5
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
